Valid state now represented by inv_bpoint==-1 (since we can have invalid state on bpoint 0)

+++ src/tsolvers/stpsolver/STPSolver.h
@@ -31,8 +31,8 @@ class STPSolver : public TSolver {
 
     std::vector<uint32_t> backtrack_points; // store of how many edges were set true at each backtrack point
 
-    size_t curr_bpoint;                     // current backtrack point
-    size_t inv_bpoint;                      // backtrack point where we entered an inconsistent state
+    uint32_t curr_bpoint;                     // current backtrack point
+    ptrdiff_t inv_bpoint;                      // backtrack point where we entered an inconsistent state
 
     ParsedPTRef parseRef(PTRef ref) const;
